+++ Models/OAuthUser.cs
@@ -7,12 +7,12 @@ namespace AspnetWeb.Models
     {
 		// UID(고유식별값), Guid, Email, mUID(User테이블)
 		[Key]
-		public int UID { get; set; }
+		public long UID { get; set; }
 
         public string GoogleUID { get; set; }
 
         public string GoogleEmail { get; set; }
 
-        public int MUID { get; set; }
+        public long MUID { get; set; }
     }
 }
